Adding godoc comments, organization.
