+++ goTezosStructs.go
@@ -5,6 +5,8 @@ import (
 	"gopkg.in/mgo.v2/bson"
 	"log"
 	"time"
+	"sync"
+	"math/rand"
 
 	"github.com/jamesruan/sodium"
 )
@@ -381,4 +383,26 @@ type Wallet struct {
 type Payment struct {
 	Address string
 	Amount float64
+}
+
+type TezClientWrapper struct {
+	healthy bool // isHealthy
+	client  *TezosRPCClient
+}
+
+/*
+ * GoTezos manages multiple Clients
+ * each Client represents a Connection to a Tezos Node
+ * GoTezos manages failover if one Node is down, there
+ * are 2 Strategies:
+ * failover: always use the same unless it is down -> go to the next - default
+ * random: send to each Node equally
+ */
+ type GoTezos struct {
+	clientLock sync.Mutex
+	RpcClients []*TezClientWrapper
+	ActiveRPCCient *TezClientWrapper
+	balancerStrategy string
+	rand *rand.Rand
+	logger *log.Logger
 }
